cakeFinallyOne

+++ src/main/java/com/irelandlight/dao/GoodsDetailDao.java
@@ -6,6 +6,7 @@ import com.irelandlight.model.Reply;
 import com.irelandlight.model.vo.*;
 import org.springframework.stereotype.Repository;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import java.util.List;
 
 /**
@@ -19,29 +20,29 @@ public interface GoodsDetailDao {
      * @param id
      * @return 返回该商品id对应的图片列表
      */
-    List<GoodsImage> queryGoodsImage(Long id);
+    List<GoodsImage> queryGoodsImage(Integer id);
 
     /**
      * 查询商品所有信息
      * @param id
      * @return 返回该商品id对应的信息
      */
-    GoodsInfo queryGoodsDetail(Long id);
+    GoodsInfo queryGoodsDetail(Integer id);
 
     /**
      * 查询该商品的价格尺寸表
      * @param id
      * @return
      */
-    List<SizePrice> querySizePrice(Long id);
+    List<SizePrice> querySizePrice(Integer id);
 
-    List<CommentGoods> queryCommentGoods(Long id);
+    List<CommentGoods> queryCommentGoods(Integer id);
     /**
      * 查询评论的所有
      * @param id
      * @return
      */
-    List<ReplyVO> queryComments(Long id);
+    List<ReplyVO> queryComments(Integer id);
     /**
      * 添加对评论的评论
      * @param reply
